remove localstorage and refactor redux
