+++ src/hoc/Sessions/WithAuthentication.js
@@ -2,23 +2,17 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { compose } from 'recompose';
 import { withFirebase } from 'src/context/firebase';
+import { setAuthUser } from 'src/modules/session';
 
 const withAuthentication = Component => {
   class WithAuthentication extends React.Component {
-    constructor(props) {
-      super(props);
-      this.props.onSetAuthUser(JSON.parse(localStorage.getItem('authUser')));
-    }
-
     componentDidMount() {
       this.listener = this.props.firebase.onAuthUserListener(
         authUser => {
           const { uid, displayName } = authUser;
-          localStorage.setItem('authUser', JSON.stringify({ uid, displayName }));
           this.props.onSetAuthUser({ uid, displayName });
         },
         () => {
-          localStorage.removeItem('authUser');
           this.props.onSetAuthUser(null);
         },
       );
@@ -34,7 +28,7 @@ const withAuthentication = Component => {
   }
 
   const mapDispatchToProps = dispatch => ({
-    onSetAuthUser: authUser => dispatch({ type: 'AUTH_USER_SET', authUser }),
+    onSetAuthUser: authUser => dispatch(setAuthUser(authUser)),
   });
 
   return compose(
