--- src/modules/reducers/session/index.js
@@ -1,20 +1,0 @@
-const INITIAL_STATE = {
-  authUser: null,
-};
-
-const applySetAuthUser = (state, action) => ({
-  ...state,
-  authUser: action.authUser,
-});
-
-function reducer(state = INITIAL_STATE, action) {
-  switch (action.type) {
-    case 'AUTH_USER_SET': {
-      return applySetAuthUser(state, action);
-    }
-    default:
-      return state;
-  }
-}
-
-export default reducer;
